moved helpers to lighten file
